+++ R/00_imports.R
@@ -1,5 +1,6 @@
 #' @import modeltime
+#' @import modeltime.resample
 #' @importFrom modeltime mdl_time_forecast modeltime_calibrate modeltime_refit mdl_time_refit get_model_description default_forecast_accuracy_metric_set summarize_accuracy_metrics
 #' @importFrom yardstick rmse rsq
-#' @importFrom tune control_resamples control_grid
+#' @importFrom tune control_grid
 NULL